Add CommandParser class
